% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{Algorithm}
\alias{Algorithm}
\title{Algorithm Class}
\description{
Class to do a non-linear minimization algorithm
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{params}}{(list) parameters solved for the model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Algorithm-new}{\code{Algorithm$new()}}
\item \href{#method-Algorithm-set_df}{\code{Algorithm$set_df()}}
\item \href{#method-Algorithm-run_nls}{\code{Algorithm$run_nls()}}
\item \href{#method-Algorithm-run_lm}{\code{Algorithm$run_lm()}}
\item \href{#method-Algorithm-solved_params}{\code{Algorithm$solved_params()}}
\item \href{#method-Algorithm-plotter}{\code{Algorithm$plotter()}}
\item \href{#method-Algorithm-plus_overlay}{\code{Algorithm$plus_overlay()}}
\item \href{#method-Algorithm-clone}{\code{Algorithm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Algorithm-new"></a>}}
\if{latex}{\out{\hypertarget{method-Algorithm-new}{}}}
\subsection{Method \code{new()}}{
Empty constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$new(passthru_df = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{passthru_df}}{(dataframe) Dataframe to pull info}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Algorithm-set_df"></a>}}
\if{latex}{\out{\hypertarget{method-Algorithm-set_df}{}}}
\subsection{Method \code{set_df()}}{
Sets the dataframe for the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$set_df(passthru_df = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{passthru_df}}{(dataframe) Dataframe to pull info}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Algorithm-run_nls"></a>}}
\if{latex}{\out{\hypertarget{method-Algorithm-run_nls}{}}}
\subsection{Method \code{run_nls()}}{
Runs the non-least squares model with passed in variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$run_nls(passthru_fncs, passthru_start)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{passthru_fncs}}{(function) For the nls to optimize}

\item{\code{passthru_start}}{(numeric vector) starting guess for nls}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self - Allows chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Algorithm-run_lm"></a>}}
\if{latex}{\out{\hypertarget{method-Algorithm-run_lm}{}}}
\subsection{Method \code{run_lm()}}{
Runs a linear model regression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$run_lm(passthru_df = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{passthru_df}}{(dataframe) Dataframe to run the lm on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self - Allows chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Algorithm-solved_params"></a>}}
\if{latex}{\out{\hypertarget{method-Algorithm-solved_params}{}}}
\subsection{Method \code{solved_params()}}{
Safley returns the parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$solved_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List - Empty if run() has not been called. Otherwise returns params
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Algorithm-plotter"></a>}}
\if{latex}{\out{\hypertarget{method-Algorithm-plotter}{}}}
\subsection{Method \code{plotter()}}{
Plots the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$plotter(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(params) Feeds into styling the plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Algorithm-plus_overlay"></a>}}
\if{latex}{\out{\hypertarget{method-Algorithm-plus_overlay}{}}}
\subsection{Method \code{plus_overlay()}}{
Function Overlay
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$plus_overlay(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(params) Feeds plot styling}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Algorithm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Algorithm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
