% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{assert}
\alias{assert}
\title{Assert Debugging Function}
\usage{
assert(passthru_msg, term = FALSE)
}
\arguments{
\item{passthru_msg}{Adds a message to the aleart}

\item{term}{Sets the function to terminating (TRUE - default) or continuing (FALSE) # nolint: line_length_linter.}
}
\value{
Void Function
}
\description{
Function alerts the user of an error and terminates
}
