% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{ggthemedefault}
\alias{ggthemedefault}
\title{ggplot2 default settings}
\usage{
ggthemedefault(passthru_gg)
}
\arguments{
\item{passthru_gg}{pass through a ggplot}
}
\description{
applys the default settings for a graph to a passed in ggplot
}
