% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{graphical_enhancment}
\alias{graphical_enhancment}
\title{Windows VScode Plot Visualization}
\usage{
graphical_enhancment()
}
\description{
alters the output of plots created in VS code to display in a serperate window, This function is windows specific
}
