% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{overlay_curve}
\alias{overlay_curve}
\title{ggplot2 Function Overlay}
\usage{
overlay_curve(passthru_plot, passthru_fncs, passthru_colour, ...)
}
\arguments{
\item{passthru_plot}{physical ggplot object}

\item{passthru_fncs}{(function/list(function)) overlayed function on the plot}

\item{passthru_colour}{(colour/ list(colour))}

\item{...}{(params - opt) passthru styling for the function overlayed}
}
\description{
Overlayed function on the plot
}
