% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{overlay_fncs}
\alias{overlay_fncs}
\title{Function Overlay}
\usage{
overlay_fncs(passthru_fncs, ...)
}
\arguments{
\item{passthru_fncs}{(function - opt) overlayed function on the plot}

\item{...}{(params - opt) passthru styling for the function overlayed}
}
\description{
Overlayed function on the plot
}
