% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{runtime_function}
\alias{runtime_function}
\title{Creates a function at run time}
\usage{
runtime_function(var_names, expression)
}
\arguments{
\item{var_names}{(list) passthrough the parameters for the new function}

\item{expression}{(quote) pass through the function body}
}
\description{
Pass through the function parameters, and the body
}
