% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{start_algorithm_nls}
\alias{start_algorithm_nls}
\title{Algorithm Class Creator}
\usage{
start_algorithm_nls(passthru_df, passthru_fncs, passthru_start)
}
\arguments{
\item{passthru_df}{(dataframe) Dataframe to pull info}

\item{passthru_fncs}{(Function) for nls to optimize}

\item{passthru_start}{(numeric vector) starting guess for nls}
}
\description{
Wraps algorithm creation to chain initialization and running the algorithm
}
