% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{start_lm}
\alias{start_lm}
\title{Linear Regression}
\usage{
start_lm(passthru_data = NA)
}
\arguments{
\item{passthru_data}{A dataframe or y~x style function for the model to analyze}
}
\value{
(list) 3 elements in order of params, std error, and the model
}
\description{
Preforms a linear regression and returns the lm object
}
