% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{start_nls}
\alias{start_nls}
\title{Non-Least Squares Regression}
\usage{
start_nls(passthru_data, passthru_fncs, passthru_start = NA)
}
\arguments{
\item{passthru_data}{A dataframe or y~x style function for the model to analyze}

\item{passthru_fncs}{the function that the template function the fit is trying to replicate}

\item{passthru_start}{starting parameters for the model to run}
}
\value{
(list) 3 elements in order of params, std error, and the model
}
\description{
Preforms a nls and returns the nls object
}
