% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winter.R
\name{start_nls_lm}
\alias{start_nls_lm}
\title{Non-Linear Regression}
\usage{
start_nls_lm(passthru_resifncs, passthru_start, ...)
}
\arguments{
\item{passthru_resifncs}{the residual function constructed between observed points and template points to minimize}

\item{passthru_start}{starting parameters for the model to run}

\item{...}{(params) MUST match the inputs of the templte function used besides optimization parameter (often in the format obs=data, var1=data1, var2=data2...)}
}
\value{
(list) 3 elements in order of params, std error, and the model
}
\description{
Preforms a non-linear regression using the e Levenberg-Marquardt algorithm and returns the nls.lm object
}
